// frontend/src/api/models/PurchaseRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PurchaseRequest = {
    name: string;
    cost: number;
    currency?: string;
    description?: (string | null);
    unit_of_measure?: string;
    project_label?: (string | null);
    is_public?: boolean;
};